Refactoring again
We now have explicit notebook files (for the .ipynb, and the associated
.py file), and a seperate python file designed to be used in code and
not in a notebook. This will be easier to edit than in jupyter.

Also pulled out some common functionality into a seperate folder, so it
can hopefully be used across algorithms.

The workflow is to interactively develop in the notebook, with all the
functions defined in line, and then pull them out and mak the seperate
file with the algorithm function, so we can use that for
comparison/testing

VPG is almost done, just want to implement GAE lambda, and then
experiment with other environments to see if it generalises.